Remove pointless abstraction layer with only one implementation
