+++ src/Offer/Security/Sapi3SearchQueryFactory.php
@@ -6,21 +6,39 @@ namespace CultuurNet\UDB3\Offer\Security;
 
 use ValueObjects\StringLiteral\StringLiteral;
 
-/**
- * Implementation of the search query factory for SAPI3.
- */
-class Sapi3SearchQueryFactory extends SearchQueryFactoryBase
+final class Sapi3SearchQueryFactory implements SearchQueryFactoryInterface
 {
-    /**
-     * @return string
-     */
-    protected function createQueryString(
+    private function createQueryString(
         StringLiteral $constraint,
         StringLiteral $offerId
-    ) {
+    ): string {
         $constraintStr = '(' . $constraint->toNative() . ')';
         $offerIdStr = $offerId->toNative();
 
         return '(' . $constraintStr . ' AND id:' . $offerIdStr . ')';
+    }
+
+    public function createFromConstraint(
+        StringLiteral $constraint,
+        StringLiteral $offerId
+    ) {
+        return new Query($this->createQueryString($constraint, $offerId));
+    }
+
+    public function createFromConstraints(
+        array $constraints,
+        StringLiteral $offerId
+    ) {
+        $queryString = '';
+
+        foreach ($constraints as $constraint) {
+            if (strlen($queryString)) {
+                $queryString .= ' OR ';
+            }
+
+            $queryString .= $this->createQueryString($constraint, $offerId);
+        }
+
+        return new Query($queryString);
     }
 }
