+++ tests/Offer/Security/Sapi3SearchQueryFactoryTest.php
@@ -4,25 +4,56 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Offer\Security;
 
+use PHPUnit\Framework\TestCase;
 use ValueObjects\StringLiteral\StringLiteral;
 
-class Sapi3SearchQueryFactoryTest extends SearchQueryFactoryTestBase
+final class Sapi3SearchQueryFactoryTest extends TestCase
 {
+    /**
+     * @var Sapi3SearchQueryFactory
+     */
+    protected $searchQueryFactory;
+
     protected function setUp()
     {
         $this->searchQueryFactory = new Sapi3SearchQueryFactory();
     }
 
     /**
-     * @return string
+     * @test
      */
-    public function createQueryString(
-        StringLiteral $constraint,
-        StringLiteral $offerId
-    ) {
-        $constraintStr = $constraint->toNative();
-        $offerIdStr = $offerId->toNative();
+    public function it_creates_a_query_from_a_constraint()
+    {
+        $constraint = new StringLiteral('zipCode:3000 OR zipCode:3010');
+        $offerId = new StringLiteral('offerId');
 
-        return '((' . $constraintStr . ') AND id:' . $offerIdStr . ')';
+        $query = $this->searchQueryFactory->createFromConstraint(
+            $constraint,
+            $offerId
+        );
+
+        $expectedQuery = new Query('((zipCode:3000 OR zipCode:3010) AND id:offerId)');
+
+        $this->assertEquals($expectedQuery, $query);
+    }
+
+    /**
+     * @test
+     */
+    public function it_creates_a_query_from_constraints()
+    {
+        $constraint1 = new StringLiteral('zipCode:3000 OR zipCode:3010');
+        $constraint2 = new StringLiteral('zipCode:3271 OR zipCode:3271');
+
+        $offerId = new StringLiteral('offerId');
+
+        $query = $this->searchQueryFactory->createFromConstraints(
+            [$constraint1, $constraint2],
+            $offerId
+        );
+
+        $expectedQuery = new Query('((zipCode:3000 OR zipCode:3010) AND id:offerId) OR ((zipCode:3271 OR zipCode:3271) AND id:offerId)');
+
+        $this->assertEquals($expectedQuery, $query);
     }
 }
